+++ src/rules/operation/RuleMintAllowance.sol
@@ -1,0 +1,60 @@
+// SPDX-License-Identifier: MPL-2.0
+pragma solidity ^0.8.20;
+
+import {AccessControlEnumerable} from "@openzeppelin/contracts/access/extensions/AccessControlEnumerable.sol";
+import {IERC165} from "@openzeppelin/contracts/utils/introspection/IERC165.sol";
+import {RuleInterfaceId} from "RuleEngine/modules/library/RuleInterfaceId.sol";
+import {ERC1404ExtendInterfaceId} from "CMTAT/library/ERC1404ExtendInterfaceId.sol";
+import {RuleEngineInterfaceId} from "CMTAT/library/RuleEngineInterfaceId.sol";
+import {IERC7551Compliance} from "CMTAT/interfaces/tokenization/draft-IERC7551.sol";
+import {IERC3643ComplianceFull} from "../../mocks/IERC3643ComplianceFull.sol";
+import {AccessControlModuleStandalone} from "../../modules/AccessControlModuleStandalone.sol";
+import {RuleMintAllowanceBase} from "./abstract/RuleMintAllowanceBase.sol";
+
+/**
+ * @title RuleMintAllowance
+ * @notice AccessControl variant of RuleMintAllowance.
+ *         `DEFAULT_ADMIN_ROLE` implicitly holds all roles.
+ *         `ALLOWANCE_OPERATOR_ROLE` can set, increase, and decrease per-minter allowances.
+ *         `COMPLIANCE_MANAGER_ROLE` can bind/unbind the rule to a RuleEngine.
+ */
+contract RuleMintAllowance is AccessControlModuleStandalone, RuleMintAllowanceBase {
+    /*//////////////////////////////////////////////////////////////
+                             CONSTRUCTOR
+    //////////////////////////////////////////////////////////////*/
+
+    /**
+     * @param admin Address of the contract admin.
+     */
+    constructor(address admin) AccessControlModuleStandalone(admin) {}
+
+    /*//////////////////////////////////////////////////////////////
+                          PUBLIC FUNCTIONS
+    //////////////////////////////////////////////////////////////*/
+
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(AccessControlEnumerable, IERC165)
+        returns (bool)
+    {
+        return interfaceId == RuleEngineInterfaceId.RULE_ENGINE_INTERFACE_ID
+            || interfaceId == ERC1404ExtendInterfaceId.ERC1404EXTEND_INTERFACE_ID
+            || interfaceId == RuleInterfaceId.IRULE_INTERFACE_ID
+            || interfaceId == type(IERC7551Compliance).interfaceId
+            || interfaceId == type(IERC3643ComplianceFull).interfaceId
+            || AccessControlEnumerable.supportsInterface(interfaceId);
+    }
+
+    /*//////////////////////////////////////////////////////////////
+                            ACCESS CONTROL
+    //////////////////////////////////////////////////////////////*/
+
+    function _authorizeSetMintAllowance() internal view virtual override onlyRole(ALLOWANCE_OPERATOR_ROLE) {}
+
+    function _onlyComplianceManager() internal virtual override onlyRole(COMPLIANCE_MANAGER_ROLE) {}
+
+    function _authorizeComplianceBindingChange(address) internal view virtual override onlyRole(COMPLIANCE_MANAGER_ROLE)
+    {}
+}